% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Quantile Regression loss as in Koenker and Bassett (1982)}
\usage{
check(x, tau)
}
\arguments{
\item{x}{Observation}

\item{tau}{Quantile probability level}
}
\description{
Quantile Regression loss as in Koenker and Bassett (1982)
}
