% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Function to check user inputs in functions. It is strongly recommended to run this before the optimizer.}
\usage{
check_inputs(
  model = NULL,
  DF,
  bounds,
  bayesopt_ctrl,
  p = 0.5,
  weights = NULL,
  seed = NULL,
  parallel = FALSE,
  score_fun = NULL,
  custom_fun = NULL,
  Y_obs = NULL,
  fit_opt_model = c("Yes", "No"),
  oos = c("No", "Yes"),
  new_data = NULL,
  honesty = TRUE,
  honesty_fraction = 0.3,
  samptype = NULL,
  splitrule = NULL,
  method = NULL,
  importance = NULL
)
}
\arguments{
\item{All}{All parameters of the optimizer}
}
\description{
Function to check user inputs in functions. It is strongly recommended to run this before the optimizer.
}
