% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_params.R
\name{opt_params}
\alias{opt_params}
\title{Extracts Optimized Parameters}
\usage{
opt_params(object)
}
\arguments{
\item{object}{optimized_model obtained after the Bayesian Optimization}
}
\description{
Extracts Optimized Parameters
}
