% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_noncrossing.R
\name{score_noncrossing}
\alias{score_noncrossing}
\title{Additional functions}
\usage{
score_noncrossing(weights, score_fun, custom_fun, DF, predictions, p, Y_obs)
}
\description{
Additional functions
}
