% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_imp.R
\name{var_imp}
\alias{var_imp}
\title{Extracts and Plot variable Importance computed with default settings of QRF models.}
\usage{
var_imp(object)
}
\arguments{
\item{object}{an optimized QRF model with ParallelBayesOptQRF or NonCrossingBayesOpt}
}
\value{
List of 'varimp_df' ordered data frame of variable importance and ggplot object 'plot_varimp'.
}
\description{
Extracts and Plot variable Importance computed with default settings of QRF models.
}
\details{
For non-crossing quantiles there is just one vector of values.
To obtain a variable importance vector for each quantile level, you should optimize each QRF separately with ParallelBayesOptQRF or NonCrossingBayesOpt.
}
